requirements.txt
django
django-rest-framework
faker
pytz
ipdb
python-decouple
whitenoise